+++ docs/changes.rst
@@ -6,6 +6,7 @@ Change Log
 
   * pysftp.Connection can be used in a `with` statement
   * add .remove() method
+  * added support for callback and confirm params to .put() method
 
 
 * 0.2.3 (released 2014-05-10)